Move memory database to api/db/memory
