+++ cityhall/test/test_env.py
@@ -13,7 +13,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from django.test import TestCase
-from lib.db.memory.cityhall_db_factory import CityHallDbFactory
+from api.db.memory.db_factory import CityHallDbFactory
 from lib.db.db import Rights
 from lib.db.connection import Connection
